[ONNX] Remove torch dependencies in _beartype
