+++ torch/onnx/_exporter_states.py
@@ -38,15 +38,3 @@ class SymbolicContext:
         self.cur_node: _C.Node = cur_node
         # Current onnx block that converted nodes are being appended to.
         self.onnx_block: _C.Block = onnx_block
-
-
-@enum.unique
-class RuntimeTypeCheckState(enum.Enum):
-    """Runtime type check state."""
-
-    # Runtime type checking is disabled.
-    DISABLED = enum.auto()
-    # Runtime type checking is enabled but warnings are shown only.
-    WARNINGS = enum.auto()
-    # Runtime type checking is enabled.
-    ERRORS = enum.auto()
