+++ torch/onnx/_internal/_beartype.py
@@ -2,14 +2,13 @@
 
 The module returns a no-op decorator when the beartype library is not installed.
 """
+import enum
 import functools
+import os
 import traceback
 import typing
 import warnings
 from types import ModuleType
-
-from torch.onnx import _exporter_states, errors
-from torch.onnx._globals import GLOBALS
 
 try:
     import beartype as _beartype_lib  # type: ignore[import]
@@ -31,17 +30,35 @@ except Exception as e:
     _beartype_lib = None  # type: ignore[assignment]
 
 
+@enum.unique
+class RuntimeTypeCheckState(enum.Enum):
+    """Runtime type check state."""
+
+    # Runtime type checking is disabled.
+    DISABLED = enum.auto()
+    # Runtime type checking is enabled but warnings are shown only.
+    WARNINGS = enum.auto()
+    # Runtime type checking is enabled.
+    ERRORS = enum.auto()
+
+
+class CallHintViolationWarning(UserWarning):
+    """Warning raised when a type hint is violated during a function call."""
+
+    pass
+
+
 def _no_op_decorator(func):
     return func
 
 
 def _create_beartype_decorator(
-    runtime_check_state: _exporter_states.RuntimeTypeCheckState,
+    runtime_check_state: RuntimeTypeCheckState,
 ):
     # beartype needs to be imported outside of the function and aliased because
     # this module overwrites the name "beartype".
 
-    if runtime_check_state == _exporter_states.RuntimeTypeCheckState.DISABLED:
+    if runtime_check_state == RuntimeTypeCheckState.DISABLED:
         return _no_op_decorator
     if _beartype_lib is None:
         # If the beartype library is not installed, return a no-op decorator
@@ -49,7 +66,7 @@ def _create_beartype_decorator(
 
     assert isinstance(_beartype_lib, ModuleType)
 
-    if runtime_check_state == _exporter_states.RuntimeTypeCheckState.ERRORS:
+    if runtime_check_state == RuntimeTypeCheckState.ERRORS:
         # Enable runtime type checking which errors on any type hint violation.
         return _beartype_lib.beartype
 
@@ -77,7 +94,7 @@ def _create_beartype_decorator(
                 # Fall back to the original function if the beartype hint is violated.
                 warnings.warn(
                     traceback.format_exc(),
-                    category=errors.CallHintViolationWarning,
+                    category=CallHintViolationWarning,
                     stacklevel=2,
                 )
 
@@ -94,6 +111,15 @@ if typing.TYPE_CHECKING:
         return func
 
 else:
-    beartype = _create_beartype_decorator(GLOBALS.runtime_type_check_state)
+    _TORCH_ONNX_EXPERIMENTAL_RUNTIME_TYPE_CHECK = os.getenv(
+        "TORCH_ONNX_EXPERIMENTAL_RUNTIME_TYPE_CHECK"
+    )
+    if _TORCH_ONNX_EXPERIMENTAL_RUNTIME_TYPE_CHECK == "WARNINGS":
+        _runtime_type_check_state = RuntimeTypeCheckState.WARNINGS
+    elif _TORCH_ONNX_EXPERIMENTAL_RUNTIME_TYPE_CHECK == "DISABLED":
+        _runtime_type_check_state = RuntimeTypeCheckState.DISABLED
+    else:
+        _runtime_type_check_state = RuntimeTypeCheckState.ERRORS
+    beartype = _create_beartype_decorator(_runtime_type_check_state)
     # Make sure that the beartype decorator is enabled whichever path we took.
     assert beartype is not None
