+++ torch/onnx/errors.py
@@ -11,7 +11,6 @@ from torch.onnx._internal import diagnostics
 __all__ = [
     "OnnxExporterError",
     "OnnxExporterWarning",
-    "CallHintViolationWarning",
     "CheckerError",
     "UnsupportedOperatorError",
     "SymbolicValueError",
@@ -20,12 +19,6 @@ __all__ = [
 
 class OnnxExporterWarning(UserWarning):
     """Base class for all warnings in the ONNX exporter."""
-
-    pass
-
-
-class CallHintViolationWarning(OnnxExporterWarning):
-    """Warning raised when a type hint is violated during a function call."""
 
     pass
